- Refactor to mimic a common ruby library - Load files through autoload - Centralize test setup through spec_helper - Remove duplicated not yet necessary spec file
